+++ src/Storage/Events/IAuditLogEvent.cs
@@ -1,0 +1,13 @@
+﻿using SenseNet.Diagnostics;
+
+// ReSharper disable once CheckNamespace
+namespace SenseNet.Events
+{
+    /// <summary>
+    /// Declares API elements for an <see cref="ISnEvent"/> that helps to write it to the Audit log.
+    /// </summary>
+    public interface IAuditLogEvent : ISnEvent
+    {
+        AuditEvent AuditEvent { get; }
+    }
+}
